requirements.txt
PyPDF2>=3.0.0
python-docx>=0.8.11
pytest>=7.0.0
playwright>=1.40.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
sqlalchemy>=2.0.23
alembic>=1.12.1
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
openai>=1.3.0
requests>=2.31.0
python-json-logger>=2.0.7

